No more absolute path in metadata cache
Reference to file are now relative to the root of the module cache
instead of being absolute.
This allows moving the full module cache around to be reused with
a different root path.
The path are also normalized, so that they are portable across systems
where the name separator changes.

Issue #1338